move fs utils, refactor tests
