+++ lib/classes/Utils.js
@@ -8,8 +8,10 @@ const fse = BbPromise.promisifyAll(require('fs-extra'));
 const _ = require('lodash');
 const fileExistsSync = require('../utils/fs/fileExistsSync');
 const writeFileSync = require('../utils/fs/writeFileSync');
+const copyDirContentsSync = require('../utils/fs/copyDirContentsSync');
 const readFileSync = require('../utils/fs/readFileSync');
 const walkDirSync = require('../utils/fs/walkDirSync');
+const dirExistsSync = require('../utils/fs/dirExistsSync');
 const isDockerContainer = require('is-docker');
 const version = require('../../package.json').version;
 const segment = require('../utils/segment');
@@ -25,12 +27,7 @@ class Utils {
   }
 
   dirExistsSync(dirPath) {
-    try {
-      const stats = fse.statSync(dirPath);
-      return stats.isDirectory();
-    } catch (e) {
-      return false;
-    }
+    return dirExistsSync(dirPath);
   }
 
   fileExistsSync(filePath) {
@@ -92,12 +89,7 @@ class Utils {
   }
 
   copyDirContentsSync(srcDir, destDir) {
-    const fullFilesPaths = this.walkDirSync(srcDir);
-
-    fullFilesPaths.forEach(fullFilePath => {
-      const relativeFilePath = fullFilePath.replace(srcDir, '');
-      fse.copySync(fullFilePath, path.join(destDir, relativeFilePath));
-    });
+    return copyDirContentsSync(srcDir, destDir);
   }
 
   generateShortId(length) {
