+++ lib/plugins/create/create.js
@@ -88,7 +88,11 @@ class Create {
     if ('template' in this.options) {
       this.createFromTemplate();
     } else if ('template-url' in this.options) {
-      return download.downloadTemplateFromRepo(this.options.name, this.options['template-url'])
+      return download.downloadTemplateFromRepo(
+        this.options['template-url'],
+        this.options.name,
+        this.options.path
+      )
         .then(dirName => {
           const message = [
             `Successfully installed "${dirName}" `,
