+++ lib/plugins/install/install.js
@@ -37,7 +37,7 @@ class Install {
   }
 
   install() {
-    return downloadTemplateFromRepo(this.options.name, this.options.url)
+    return downloadTemplateFromRepo(this.options.url, this.options.name)
       .then(dirName => {
         const message = [
           `Successfully installed "${dirName}" `,
