+++ lib/utils/downloadTemplateFromRepo.js
@@ -9,26 +9,13 @@ const qs = require('querystring');
 
 const { renameService } = require('./renameService');
 const { ServerlessError } = require('../classes/Error');
-const walkDirSync = require('./fs/walkDirSync');
+const copyDirContentsSync = require('./fs/copyDirContentsSync');
+const dirExistsSync = require('./fs/dirExistsSync');
 
-function copyDirContentsSync(srcDir, destDir) {
-  const fullFilesPaths = walkDirSync(srcDir);
-
-  fullFilesPaths.forEach(fullFilePath => {
-    const relativeFilePath = fullFilePath.replace(srcDir, '');
-    fse.copySync(fullFilePath, path.join(destDir, relativeFilePath));
-  });
-}
-
-function dirExistsSync(dirPath) {
-  try {
-    const stats = fse.statSync(dirPath);
-    return stats.isDirectory();
-  } catch (e) {
-    return false;
-  }
-}
-
+/**
+ * @param {Object} url
+ * @returns {Object}
+ */
 function parseGitHubURL(url) {
   const parts = url.pathname.split('/');
   const isSubdirectory = parts.length > 4;
@@ -70,6 +57,10 @@ function parseGitHubURL(url) {
   };
 }
 
+/**
+ * @param {Object} url
+ * @returns {Object}
+ */
 function parseBitBucketURL(url) {
   const parts = url.pathname.split('/');
   const isSubdirectory = parts.length > 4;
@@ -113,6 +104,13 @@ function parseBitBucketURL(url) {
   };
 }
 
+/**
+ * Parse URL and call the appropriate adaptor
+ *
+ * @param {string} inputUrl
+ * @throws {ServerlessError}
+ * @returns {Object}
+ */
 function parseRepoURL(inputUrl) {
   if (!inputUrl) {
     throw new ServerlessError('URL is required');
@@ -139,7 +137,13 @@ function parseRepoURL(inputUrl) {
   }
 }
 
-function downloadTemplateFromRepo(name, inputUrl) {
+/**
+ * @param {string} inputUrl
+ * @param {string} [templateName]
+ * @param {string} [path]
+ * @returns {Promise}
+ */
+function downloadTemplateFromRepo(inputUrl, templateName, downloadPath) {
   const repoInformation = parseRepoURL(inputUrl);
 
   let serviceName;
@@ -149,11 +153,11 @@ function downloadTemplateFromRepo(name, inputUrl) {
   if (repoInformation.isSubdirectory) {
     const folderName = repoInformation.pathToDirectory.split('/').splice(-1)[0];
     serviceName = folderName;
-    dirName = name || folderName;
+    dirName = downloadPath || templateName || folderName;
     downloadServicePath = path.join(os.tmpdir(), repoInformation.repo);
   } else {
     serviceName = repoInformation.repo;
-    dirName = name || repoInformation.repo;
+    dirName = downloadPath || templateName || repoInformation.repo;
     downloadServicePath = path.join(process.cwd(), dirName);
   }
