+++ lib/utils/downloadTemplateFromRepo.test.js
@@ -47,24 +47,24 @@ describe('downloadTemplateFromRepo', () => {
 
   describe('downloadTemplateFromRepo', () => {
     it('should throw an error if the passed URL option is not a valid URL', () => {
-      expect(() => downloadTemplateFromRepo(null, 'invalidUrl')).to.throw(Error);
+      expect(() => downloadTemplateFromRepo('invalidUrl')).to.throw(Error);
     });
 
     it('should throw an error if the passed URL is not a valid GitHub URL', () => {
-      expect(() => downloadTemplateFromRepo(null, 'http://no-github-url.com/foo/bar')).to.throw(Error);
+      expect(() => downloadTemplateFromRepo('http://no-github-url.com/foo/bar')).to.throw(Error);
     });
 
     it('should throw an error if a directory with the same service name is already present', () => {
       const serviceDirName = path.join(servicePath, 'existing-service');
       fse.mkdirsSync(serviceDirName);
 
-      expect(() => downloadTemplateFromRepo(null, 'https://github.com/johndoe/existing-service')).to.throw(Error);
+      expect(() => downloadTemplateFromRepo('https://github.com/johndoe/existing-service')).to.throw(Error);
     });
 
     it('should download the service based on the GitHub URL', () => {
       const url = 'https://github.com/johndoe/service-to-be-downloaded';
 
-      return downloadTemplateFromRepo(null, url).then(() => {
+      return downloadTemplateFromRepo(url).then(() => {
         expect(downloadStub.calledOnce).to.equal(true);
         expect(downloadStub.args[0][0]).to.equal(`${url}/archive/master.zip`);
       });
@@ -92,7 +92,7 @@ describe('downloadTemplateFromRepo', () => {
             writeFileSync(slsYml, 'service: service-name');
           }));
 
-      return downloadTemplateFromRepo(name, url).then(() => {
+      return downloadTemplateFromRepo(url, name).then(() => {
         expect(downloadStub.calledOnce).to.equal(true);
         expect(downloadStub.args[0][1]).to.contain(name);
         expect(downloadStub.args[0][0]).to.equal(`${url}/archive/master.zip`);
@@ -116,7 +116,7 @@ describe('downloadTemplateFromRepo', () => {
             writeFileSync(slsYml, 'service: service-name');
           }));
 
-      return downloadTemplateFromRepo(name, url).then(() => {
+      return downloadTemplateFromRepo(url, name).then(() => {
         expect(downloadStub.calledOnce).to.equal(true);
         const yml = readFileSync(path.join(newServicePath, 'serverless.yml'));
         expect(yml.service).to.equal(name);
